Cleanup

+++ ginvoice/ui/record.py
@@ -65,7 +65,7 @@ class RecordDialog(Gtk.Window):
             record.quantity_postfix = _('h')
         elif qtype == 2:
             record.quantity_postfix = _('m')
-        record.quantity = int(self.quantity.get_text()) if qtype == 0 else float(self.quantity.get_text())
+        record.quantity = float(self.quantity.get_text()) if qtype == 0 else float(self.quantity.get_text())
         record.price = float(self.price.get_text())
         if self.percentages_radio.get_active():
             record.discount = record.quantity * record.price * (float(self.discount.get_text()) / 100)
